requirements.txt
streamlit
pandas
plotly
python-dotenv
anthropic
requests
sqlalchemy
pydantic
passlib
bcrypt
